requirements.txt
fastapi
uvicorn
websockets
httpx
python-dotenv
python-multipart
jinja2
starlette
certifi
pyinstaller